전체 점검: 패턴 타공 표시선 수집 결함 등 수정 + 설치본 v62
정적 전면 검토로 찾은 실제/잠재 결함을 고쳤다. 상세는 docs/pattern-studio,
docs/quick-toggle, docs/warning-pick, docs/model-sync, docs/wall-floor-split의
2026-09-01 항목 참고.

패턴 (요청 중점 영역)
- PatternDisplayedLineCollector가 호스트 문서를 Document.Equals(참조 비교)로
  판정해, false가 되면 내보낸 선을 전부 버려 패턴 타공이 항상 "표시 패턴 선을
  찾지 못했습니다"로 끝날 수 있었다. 경고Pick의 CONFIRMED LIVE BUG와 같은 원인.
  경로/제목 기반 DocumentKey 값 비교 + 링크 여부 판정으로 교체.
- ModelLinePatternCaptureCommand.RationalDirection: 거의 수직인 캡처 선에서
  (int)Math.Round 오버플로 -> Math.Abs(int.MinValue) OverflowException.
- PatternStudioWindow.BuildGridGeometry / PatternLineGenerator.Generate:
  반복 인덱스 한도 검사의 int 오버플로로 수십억 회 루프(Revit 정지) 가능.
  long 산술 + Offset 0 가드.
- PatternStudioWindow.SetSelectedGridControls의 _suppressUi 재진입 가드.
- ExecuteWithNewWallSketch가 probe를 항상 false로 넘겨 사전 검증 중에도 실제
  적용 경로가 돌고 복원 기록이 중복되던 문제 (PunchExecutionResult.RecordAppended).
- PatternTransformService 중복 인덱스 예외, PatternPunchRecordStore.Read 가드,
  PAT 헤더 앞 %TYPE 인식, 모델선 캡처 덮어쓰기 성공 문구.

그 밖
- Revit Idling/ViewActivated/DocumentClosing과 IExternalEventHandler.Execute의
  예외 차단(콜백 예외는 대화상자 폭주 또는 "반응 없음"으로만 남는다).
- WPF 핸들러 안 설정/기록 저장 예외 차단, ChangeLog.Append 저장 실패 무시.
- ModelSyncWindow의 Document ReferenceEquals 비교 제거.
- SplitWallCommand 널 가드 3곳, SunnyToolsCommands.All에 경고Pick 추가.

검증: Revit 2023~2027 Release 빌드 오류 0개(경고 36/56/58/38/40, 5개씩 감소).
Payload 28개 DLL이 빌드 원본과 SHA-256 일치, 5개 연도 DLL에서 수정 심볼 확인,
모든 연도 Clipper2Lib 포함/RevitAPI 미포함, 게시->ZIP 30개 파일 해시 일치.
라이브 검증은 아직 없음 - Revit 종료 후 v62 설치해 패턴 타공부터 확인할 것.

+++ WallSplitter/PatternGeometry.cs
@@ -119,11 +119,13 @@ namespace WallSplitter
                 double maxProjection = corners.Max(point => Dot(point, normal));
                 double first = (minProjection - originProjection) / offset;
                 double last = (maxProjection - originProjection) / offset;
-                int start = SafeFloor(Math.Min(first, last)) - 2;
-                int end = SafeCeiling(Math.Max(first, last)) + 2;
+                // 아주 작은 Offset 때문에 first/last가 int 범위를 넘으면 클램프된 두 값의 int 뺄셈이
+                // 오버플로해 아래 한도 검사를 그냥 통과한다(수십억 회 루프 = Revit 정지). long으로 비교한다.
+                long start = (long)SafeFloor(Math.Min(first, last)) - 2;
+                long end = (long)SafeCeiling(Math.Max(first, last)) + 2;
                 if (end - start + 1 > maximumLines)
                 {
-                    int center = SafeFloor((first + last) * 0.5);
+                    long center = SafeFloor((first + last) * 0.5);
                     start = center - maximumLines / 2;
                     end = center + maximumLines / 2;
                     result.WasLimited = true;
@@ -131,7 +133,7 @@ namespace WallSplitter
                 }
 
                 List<double> scaledSegments = source.Segments.Select(value => value * lengthScale).ToList();
-                for (int k = start; k <= end; k++)
+                for (long k = start; k <= end; k++)
                 {
                     PatternPoint anchor = origin + (direction * shift + normal * offset) * k;
                     if (!TryClipInfiniteLine(anchor, direction, bounds, out double t0, out double t1)) continue;
